Sns mod for sns

+++ MiraAPI.Example/ExamplePlugin.cs
@@ -1,6 +1,7 @@
 using HarmonyLib;
+using MiraAPI;
 using MiraAPI.Options;
-using MiraAPI.Roles;
+using System;
 
 namespace MiraAPI.Example
 {
@@ -12,11 +13,14 @@ namespace MiraAPI.Example
         public override void Load()
         {
             var options = GameOptionsManager.Instance.currentVars;
-            options.SabotageCooldown = 20f; 
-            options.ReportDistance = 0f;
-            options.NumEmergencyMeetings = 0;
-            options.NumShapeshifters = 3;
-            options.ShapeshifterChance = 100;
+            if (options != null)
+            {
+                options.SabotageCooldown = 20f; 
+                options.ReportDistance = 0f;
+                options.NumEmergencyMeetings = 0;
+                options.NumShapeshifters = 3;
+                options.ShapeshifterChance = 100;
+            }
             Harmony.PatchAll();
         }
     }
@@ -27,7 +31,7 @@ namespace MiraAPI.Example
         [HarmonyPostfix]
         public static void Postfix(ChatController __instance)
         {
-            if (__instance.TextArea.text.ToLower().Contains("/r"))
+            if (__instance != null && __instance.TextArea.text.ToLower().Contains("/r"))
             {
                 __instance.AddChat(PlayerControl.LocalPlayer, "SNS: Shift-kill only. Comms only. No reports.");
             }
